+++ src/assets.zig
@@ -110,7 +110,7 @@ pub const Assets = struct {
             .laser => getLaserSprite(ship),
             .storage => getStorageSprite(ship),
             .railgun => getRailgunSprite(ship),
-            // else => Sprite.initEmpty(),
+            .smart_core => getSmartCoreSprite(ship),
         };
     }
 
@@ -167,5 +167,15 @@ pub const Assets = struct {
         }
 
         return Sprite.init(ship_sheet, setRow(5, index));
+    }
+
+    pub fn getSmartCoreSprite(ship: ShipPartTileType) Sprite {
+        var index = @intFromEnum(ship.rotation);
+
+        if (PartStats.isBroken(ship)) {
+            index += 1;
+        }
+
+        return Sprite.init(ship_sheet, setRow(6, index));
     }
 };